main/backend/requirements.txt
livekit
livekit-agents[openai]
livekit-plugins-openai
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
pypdf
openai